[PROJECT] Added play progres monitor.
A module with some classes and functions to provide an easy way for replay channels to monitor a stream and report the play position to the web service at regular intervals.
All to support 2-way sync of features like continue watching, so user can continue watching on other devices what they started on Kodi.